backend-python/requirements.txt
# AXS360 - Production Backend API
# FastAPI + PostgreSQL + Redis + Stripe + Twilio

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.0.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
celery==5.3.4
stripe==7.5.0
twilio==8.10.0
sendgrid==6.10.0
qrcode[pil]==7.4.2
pillow==10.0.1
reportlab==4.0.7
pandas==2.1.3
numpy==1.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-dotenv==1.0.0
gunicorn==21.2.0
httptools==0.6.1
websockets==12.0
